http3: move WebTransport interfaces to interface.go
